pyproject.toml
[tool.poetry]
name = "autoria-clone"
version = "0.1.0"
description = ""
authors = ["AlexJohannson <mehedsanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
mysqlclient = "^2.2.7"
djangorestframework-simplejwt = "^5.5.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
